fix: cast scaled font point sizes to int in FormatText
wxPython Phoenix's Font.SetPointSize rejects float; fontSize*<factor>
produced a float, crashing FormatText on compare/heading styles.

+++ read/view.py
@@ -474,22 +474,22 @@ class View(AuiBaseFrame):
                 font.SetPointSize(fontSize)
                 readPanel.Body.SetStyle(int(x), int(y), wx.TextAttr('blue', wx.NullColour, font))  
             elif tag == 'eh1':
-                font.SetPointSize(fontSize*1.2)
+                font.SetPointSize(int(fontSize*1.2))
                 readPanel.Body.SetStyle(int(x), int(y), 
                     wx.TextAttr(utils.LoadThemeForegroundHex(constants.READ), wx.NullColour, font))  
             elif tag == 'eh2':
-                font.SetPointSize(fontSize*0.85)                
+                font.SetPointSize(int(fontSize*0.85))
                 readPanel.Body.SetStyle(int(x), int(y), 
                     wx.TextAttr(wx.Colour(0x88,0x88,0x88,0xFF), wx.NullColour, font))  
             elif tag == 'eh3':
-                font.SetPointSize(fontSize*0.75)
+                font.SetPointSize(int(fontSize*0.75))
                 readPanel.Body.SetStyle(int(x), int(y), 
                     wx.TextAttr(wx.Colour(0x88,0x88,0x88,0xFF), wx.NullColour, font))  
             elif tag == 'er':
-                font.SetPointSize(fontSize*0.70)
+                font.SetPointSize(int(fontSize*0.70))
                 readPanel.Body.SetStyle(int(x), int(y), wx.TextAttr('#919394', wx.NullColour, font))
             elif tag == 'fn':
-                font.SetPointSize(fontSize*0.8)
+                font.SetPointSize(int(fontSize*0.8))
                 readPanel.Body.SetStyle(int(x), int(y), wx.TextAttr('#3CBF3F', wx.NullColour, font))  
         
     def ShowFindDialog(self, code, index, text, flags):
